Create Exercise 8
Look for a command from the numpy documentation (or elsewhere on the Internet) to calculate the matrix determinant. The determinant is a number that does not require a deeper understanding here. Calculate the determinants of the matrices A, B and AB, where

A=\begin{pmatrix}
1&2\\
3&4
\end{pmatrix}{,} B=\begin{pmatrix}
-1  & 1\\
5 & 7
\end{pmatrix}

and AB stands for matrix product. State that det (AB) = det (A) det (B) can be understood as the product of the numbers (the rounding error can be ignored).